Worked with ChatGPT on many approaches to parse RDF using XmlReader and deserialize it to an object model suggested by ChatGPT. The code is perfectly fine after small adjustments and refinements. The issue is that XmlReader appears unable to accomodate xml namespaces in an RDF document and generates an exception.
Multiple variations of the code was made with and without consulting
ChatGPT and the results all produced the following exception.

Inner Exception 1:
InvalidOperationException: <RDF xmlns='http://www.w3.org/1999/02/22-rdf-syntax-ns#'> was not expected.

System.InvalidOperationException
  HResult=0x80131509
  Message=There is an error in XML document (2, 2).
  Source=System.Private.Xml
  StackTrace:
   at System.Xml.Serialization.XmlSerializer.Deserialize(XmlReader xmlReader, String encodingStyle, XmlDeserializationEvents events)
   at System.Xml.Serialization.XmlSerializer.Deserialize(XmlReader xmlReader)
   at gautier.rss.data.RDFConversion.RDFConverter.CreateRDF(String rdfFilePath)

It is possible that the source data is invalid but since it is data from
a third-party, the responsibility to adjust is on the software. For now,
I am going to use the less efficient XmlDocument from .NET 1.0 to load
the document, process it and address the issue efficient
deserialization at a future time.