In heavy collaboration with ChatGPT, produced a data model in C# for a reference RDF document from https://distrowatch.com/news/dw.xml. I then guided ChatGPT to produce a process in C# to load process the RDF document into the data model and then convert the data model to an instance of SyndicationFeed.
This required multiple revision iterations that combined the
capabilities of ChatGPT with years of personal experience to derive a
solution that compiled properly with reduced defects.

This solution supercedes the prior approach that used an XmlTextReader
with one that uses XmlDocument. The reason for this shift is XmlDocument
is more flexible when parsing the data at the cost of higher memory use.

The performance side-effects will be addressed in the future. It was
more important to move towards a completed process over upfront
optimization in this case.