Redefined from a readonly struct to read/writeable class definition. This was done to favor flexibility in the current implementation wave. Should a readonly struct record type be desired in the future, the approach will go towards a wrapper that will convert instances of this type or this type's input data stream to the corresponding record type if necessary.
It was also more expedient to transform this type definition this way
than change new implementation code to correspond to the constraints of
the prior class/struct definition.

The balance of downstream changes to other parts of the implementation
was considered less substantial compared with preserving the semantics
of newly added implementation code.