Centralize fundamental RSS communication and API representation. I want the RSSNetClient to handle the core interaction with network sources of RSS data. This positions RSSNetClient as the key point in the system to ensure reliable access and mitigation of RSS data download.
I want RSS data initially captured here and the first decision on how
that data [in its basic form] is represented to the rest of the system defined here.

Basic RSS format is represented 3-fold:

- Raw Xml data - In-Memory.
- Raw Xml data - Persist to File.
- Microsoft Syndication Feed.

Raw Xml data is typically accessed through an http or https endpoint.

The current implementation does not maintain an in-memory version of the
data longer than necessary to produce an file that mirrors the in-memory
version.

The Microsoft Syndication Feed is a 'convenience' representation used to
reduce the amount of explicit software code instructions needed at the
application level to convert from Raw Xml version of RSS data to the
Gautier RSS Feed API.

Centralizing these functions in this static class provides fundamental
control over the sequence and error handling involved in the handshake
from http(s) to Gautier RSS Feed API.