Compare file date/time between 2 files.
Files in the RSS process defined in this project depend on each other. When an RSS feed is downloaded to a .xml file, that file is used to create a .txt file. The .txt file is in tab-delimited format and used to update the database. Comparing the date/time between these files stops the program from generating a new copy of the tab-delimited files when the rss xml file did not change.

Temporal evaluation of the files eliminates repetative file generation. Integration processing is controlled at the same time. Detection of the valid state of existing files blocks repetative database updates. Moderating the updates using the date/time evaluation reduces redundant update activity and boosts performance.